Remove extra public api
